fix schema build
